+++ src/engraving/libmscore/beam.h
@@ -114,7 +114,7 @@ public:
 
     // Score Tree functions
     EngravingObject* scanParent() const override;
-    EngravingObject* scanChild(int idx) const override;
+    EngravingObject* scanChild(size_t idx) const override;
     size_t scanChildCount() const override;
 
     Beam* clone() const override { return new Beam(*this); }
