+++ src/engraving/libmscore/measurebase.h
@@ -99,7 +99,7 @@ public:
 
     // Score Tree functions
     EngravingObject* scanParent() const override;
-    EngravingObject* scanChild(int idx) const override;
+    EngravingObject* scanChild(size_t idx) const override;
     size_t scanChildCount() const override;
     virtual void scanElements(void* data, void (* func)(void*, EngravingItem*), bool all=true) override;
